Refactor notifications and distinguish better internal vs system.
